+++ app/routes/components/ScrollArea/changelog.codegen.ts
@@ -10,4 +10,5 @@ export const changelog = ''
   + '  alpha when dragging.\n'
   + '- 2025-10-30: Add the variant `absolute` (it uses absolute positioning to make\n'
   + '  viewport just fill the container).\n'
+  + '- 2025-11-04: Support the `ScrollArea.ScrollbarWithThumb`.\n'
   ;