+++ src/components/ScrollArea/index.ts
@@ -1,3 +1,3 @@
-// ScrollArea from copy-ui @ 2025-10-30.
+// ScrollArea from copy-ui @ 2025-11-04.
 
 export { ScrollArea } from "./scroll-area";